requirements.txt
Flask==2.1.1
google-cloud-pubsub==2.9.0
google-cloud-bigquery==2.29.0
unittest2==1.1.0
werkzeug==2.0.3